Implemented proper view frustum
Well, it's not a frustrum since it doesn't have a close or far plane, but it serves the same function. The cleaner solution to out-of-screen triangles reduces code size and visual bugs, but is slower. The visual bugs can hopefully be improved further, and I think the performance can be much, much better