Added near plane for portal view
Also cleaned up a bit. Now the portal view is almost perfect!

If an object is partway through a portal, then the clone on the farside will be bisected by the near plane, and the one on this side will be partially covered by the portal. This makes it slightly less than perfect